pyproject.toml
[tool.poetry]
name = "yura-bot"
version = "0.1.0"
description = "Discord bot for Twitter notifications"
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.11.0"
tweety-ns = "1.0.9"
discord-py = "2.3.2"
pydantic = "2.5.3"
pydantic-settings = "2.1.0"
pyyaml = "6.0.1"
loguru = "0.7.2"
boto3 = "1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-asyncio = "0.21.1"
black = "23.12.1"
ruff = "0.1.9"
mypy = "1.8.0"
pre-commit = "3.6.0"

[tool.poetry.group.infra.dependencies]
pulumi = "3.100.0"
pulumi-aws = "6.18.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
